Updating README and adding twitter and github to config. Renamed components
