+++ src/components/Post/Post.css
@@ -1,5 +1,5 @@
 @media all and (min-width: 600px) {
-  .appbox {
+  .post {
     width: 80%;
     margin: 25px;
     padding-left: 25px;
@@ -13,7 +13,7 @@
   }
 }
 @media all and (max-width: 600px) {
-  .appbox {
+  .post {
     width: 80%;
     margin: 25px;
     padding-left: 5px;
