+++ src/components/Posts/Posts.css
@@ -1,4 +1,4 @@
-.Apps {
+.Posts {
   margin: 0 auto;
   max-width: 650px;
 }
