+++ src/pages/Courses/Courses.js
@@ -1,6 +1,5 @@
-import blogs from "../../constants/courses";
 import React from "react";
-import AppList from "../../components/Apps/Apps";
+import Posts from "../../components/Posts/Posts";
 import Header from "../../components/Header/Header";
 import courses from "../../constants/courses";
 import "./Courses.css";
@@ -16,7 +15,7 @@ export default function(props) {
   return (
     <div>
       <Header path={props.location.pathname} />
-      <AppList list={courses} action={courseLink} />
+      <Posts list={courses} action={courseLink} />
     </div>
   );
 }
